Fixed restoration of the panel divider position.
Moved the restoration of the divider position in the ui_window_show
function. The ui_window_update function does not be called
after closing the preference window or in check_ui_visible, it
was causing a restoration of the divider position to the old
setting.

Based on patch of LP #1642034 (Francis Chin)